Move overlapping encrypted paths test into encryption.test.ts
Remove the exploratory probe file and keep only the single test that
actually reproduces the bug, adding it to the ISSUES section of the
existing encryption test suite.
