+++ test/unit.test.ts
@@ -58,7 +58,6 @@ import './unit/replication-graphql.test.ts';
 import './unit/replication-websocket.test.ts';
 import './unit/replication-webrtc.test.ts';
 import './unit/encryption.test.ts';
-import './unit/encryption-bug-probe.ts';
 import './unit/rx-state.test.ts';
 import './unit/migration-schema.test.ts';
 import './unit/attachments.test.ts';
